+++ src/main/scala/edu/ie3/simona/agent/participant/ParticipantAgentFundamentals.scala
@@ -49,7 +49,7 @@ import edu.ie3.simona.agent.state.ParticipantAgentState.{
   Calculate,
   HandleInformation,
 }
-import edu.ie3.simona.config.SimonaConfig
+import edu.ie3.simona.config.RuntimeConfig.BaseRuntimeConfig
 import edu.ie3.simona.event.ResultEvent
 import edu.ie3.simona.event.ResultEvent.{
   FlexOptionsResultEvent,
@@ -108,7 +108,7 @@ protected trait ParticipantAgentFundamentals[
     MS <: ModelState,
     D <: ParticipantStateData[PD],
     I <: SystemParticipantInput,
-    MC <: SimonaConfig.BaseRuntimeConfig,
+    MC <: BaseRuntimeConfig,
     M <: SystemParticipant[CD, PD, MS],
 ] extends ServiceRegistration[PD, CD, MS, D, I, MC, M] {
   this: ParticipantAgent[PD, CD, MS, D, I, MC, M] =>